Rename AbstractApiTest to BaseControllerTestCase.php
