+++ tests/Controller/API/BaseControllerTestCase.php
@@ -5,7 +5,7 @@ namespace App\Tests\Controller\API;
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
-abstract class AbstractApiTest extends WebTestCase
+abstract class BaseControllerTestCase extends WebTestCase
 {
     protected $client;
