+++ tests/Controller/API/DomainApiTestCase.php
@@ -2,11 +2,11 @@
 
 namespace App\Tests\Controller\API;
 
-class AlertDestinationApiTest extends AbstractApiTest
+class DomainApiTestCase extends BaseControllerTestCase
 {
     public function testCollection()
     {
-        $this->client->request('GET', '/api/alert_destinations.json', [], [], [
+        $crawler = $this->client->request('GET', '/api/domains.json', [], [], [
             'HTTP_Accept' => 'application/json',
         ]);
 
@@ -18,18 +18,16 @@ class AlertDestinationApiTest extends AbstractApiTest
 
     public function testAddRemove()
     {
-        $this->client->request(
+        $crawler = $this->client->request(
             'POST',
-            '/api/alert_destinations.json',
+            '/api/domains.json',
             [],
             [],
             [
                 'CONTENT_TYPE' => 'application/json',
             ],
             json_encode([
-                'name' => 'syslogtest',
-                'type' => 'syslog',
-                'parameters' => [],
+                'name' => 'New Domain',
             ])
         );
 
@@ -42,10 +40,24 @@ class AlertDestinationApiTest extends AbstractApiTest
 
         $crawler = $this->client->request(
             'DELETE',
-            "/api/alert_destinations/$id.json"
+            "/api/domains/$id.json"
         );
 
         $response = $this->client->getResponse();
         $this->assertEquals(204, $response->getStatusCode());
+    }
+
+    public function testAlerts()
+    {
+        $crawler = $this->client->request('GET', '/api/domains/1/alerts.json', [], [], [
+            'HTTP_Accept' => 'application/json',
+        ]);
+
+        $response = $this->client->getResponse();
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertTrue($response->headers->contains('Content-Type', 'application/json'));
+        $this->assertJson($response->getContent());
+
+        $this->assertEquals('Alertrule 1 on Device 1 from Slavegroup 1', json_decode($response->getContent())[0]->message);
     }
 }
